requirements.txt
langchain==0.0.128
openai==0.27.2
faiss-cpu==1.7.3
flask==2.2.3
flask-socketio==5.3.3
flask-executor==1.0.0
gevent==22.10.2
gevent-websocket==0.10.1
unstructured==0.5.8
flask-cors==3.0.10